requirements.txt
django==4.0.4
djangorestframework==3.13.1
environ
psycopg2-binary==2.9.3
